requirements.txt
FPDF==1.7.2
numpy==1.25.1
pandas==2.0.3
matplotlib==3.7.2
python-bidi==0.4.2
arabic-reshaper==3.0.0
pyyaml==6.0.1
